修复收集器行数计算规则

+++ option.go
@@ -28,7 +28,7 @@ type (
 		notifiers []notifier
 
 		pwe bool
-		max int64
+		max int
 	}
 )
